correcoes

+++ www/listar.php
@@ -17,24 +17,31 @@
 
    <?php 
   
-   $tabela = json_decode($curl->getLista(false)->getResultado());
-   foreach($tabela as $chave => $valor){
+// try {
+   
+ 
+   if($tabela = json_decode($curl->getLista(false)->getResultado())){
 
-      echo "<tr><th scope='row'>{$valor->codigo}</th>";
-    
-      foreach($valor as $chaves => $valor2){
-        if($chaves == "codigo")
-          continue;
+     foreach($tabela as $chave => $valor){
 
-        echo "<td>".(substr($chaves,0,4)=="data" ? $funcoes::corrigirData($valor2) : $valor2)."</td>";
-    
-      }
+        echo "<tr><th scope='row'>{$valor->codigo}</th>";
+      
+        foreach($valor as $chaves => $valor2){
 
-    echo '<td><a href=?pag=alterarRegistro&codigo='.$valor->codigo.'><span class="glyphicon glyphicon-pencil"></span>Alterar</a></td>';
-    echo "<td><a href=?pag=apagarRegistro&codigo=".$valor->codigo." onclick=\"return confirm('Deseja Realmente Excluir?')\">Excluir</a></td>";
-    echo "</tr>";
+          if($chaves == "codigo") continue;
+          echo "<td>".(substr($chaves,0,4)=="data" ? $funcoes::corrigirData($valor2) : $valor2)."</td>";
+      
+        }
 
-  }
+      echo '<td><a href=?pag=alterarRegistro&codigo='.$valor->codigo.'><span class="glyphicon glyphicon-pencil"></span>Alterar</a></td>';
+      echo "<td><a href=?pag=apagarRegistro&codigo=".$valor->codigo." onclick=\"return confirm('Deseja Realmente Excluir?')\">Excluir</a></td>";
+      echo "</tr>";
+
+    }
+
+}else{
+    echo '<tr><td colspan=9 align="center">Tabela Vazia!!</td></tr>';
+  }  
 
   ?>
  </tbody>
